speakeasy/requirements.txt
# =============================================
# META DATA HEADER
# Name: requirements.txt - SPEAKEASY Dependencies
# Date: 2026-02-11
# Version: 0.1.0
# Category: SPEAKEASY
#
# CHANGELOG:
#   - v0.1.0 (2026-02-11): Initial template
#     * Python 3.12+ required
#     * Dependencies listed as-needed
# =============================================

# AIPass SPEAKEASY - Python Dependencies
# Python >=3.12 required
#
# Installation:
#   pip install -r requirements.txt
#
# For development dependencies:
#   pip install -e ".[dev]"

# Core Dependencies

# Audio Processing
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.24.3

# Speech Recognition
faster-whisper>=1.0.2
webrtcvad-wheels>=2.0.11

# Input/Output
pynput>=1.7.6
PyQt5>=5.15.10

# Configuration
PyYAML>=6.0.1

# Utilities
rich>=13.0.0

# Development Dependencies (optional, can also use pyproject.toml [project.optional-dependencies])
# pytest>=7.4.3
# pytest-cov>=4.1.0
